Remove LifecycleStateMachine in favor of extension functions
